+++ bootique-aws2-s3-junit5/src/main/java/io/bootique/aws2/s3/junit5/S3TestFactory.java
@@ -22,41 +22,41 @@ import com.fasterxml.jackson.annotation.JsonTypeName;
 import io.bootique.annotation.BQConfig;
 import io.bootique.aws2.AwsConfig;
 import io.bootique.aws2.junit5.AwsTester;
-import io.bootique.aws2.s3.S3Factory;
+import io.bootique.aws2.s3.S3ClientFactory;
+import io.bootique.aws2.s3.S3ClientFactoryFactory;
 import io.bootique.di.Injector;
 import io.bootique.di.Key;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import software.amazon.awssdk.services.s3.S3Client;
 
 @BQConfig
 @JsonTypeName("awss3test")
-public class S3TestFactory extends S3Factory {
+public class S3TestFactory extends S3ClientFactoryFactory {
 
     static final Logger LOGGER = LoggerFactory.getLogger(S3TestFactory.class);
 
     @Override
-    public S3Client createS3(AwsConfig config, Injector injector) {
+    public S3ClientFactory create(AwsConfig config, Injector injector) {
         Key<AwsTester> awsTesterKey = Key.get(AwsTester.class);
         if (!injector.hasProvider(awsTesterKey)) {
             LOGGER.warn("Using S3Tester without AwsTester. This will likely result in an invalid configuration");
-            return super.createS3(config, injector);
+            return super.create(config, injector);
         }
 
         AwsTester tester = injector.getInstance(awsTesterKey);
         setEndpointOverride(tester.getEndpointOverride());
-        S3Client client = super.createS3(config, injector);
+        S3ClientFactory factory = super.create(config, injector);
 
         // run client startup callbacks
         Key<S3Tester> s3TesterKey = Key.get(S3Tester.class);
         if (!injector.hasProvider(s3TesterKey)) {
             LOGGER.warn("No S3Tester in the environment. Skipping tester initialization");
-            return client;
+            return factory;
         }
 
         S3Tester s3Tester = injector.getInstance(s3TesterKey);
-        s3Tester.onS3ClientInit(client);
+        s3Tester.onS3ClientFactoryInit(factory);
 
-        return client;
+        return factory;
     }
 }