+++ bootique-aws2-s3-junit5/src/test/java/io/bootique/aws2/s3/junit5/S3Tester_CreateBucketsIT.java
@@ -20,6 +20,7 @@ package io.bootique.aws2.s3.junit5;
 
 import io.bootique.BQRuntime;
 import io.bootique.Bootique;
+import io.bootique.aws2.s3.S3ClientFactory;
 import io.bootique.junit5.BQApp;
 import io.bootique.junit5.BQTest;
 import io.bootique.junit5.BQTestTool;
@@ -49,7 +50,7 @@ public class S3Tester_CreateBucketsIT extends BaseAwsTest {
 
     @Test
     public void test() throws IOException {
-        S3Client s3 = app.getInstance(S3Client.class);
+        S3Client s3 = app.getInstance(S3ClientFactory.class).newClient();
 
         s3.putObject(b -> b.bucket("bt1").key("b1k1"), RequestBody.fromBytes("b1v1".getBytes()));
         s3.putObject(b -> b.bucket("bt2").key("b2k1"), RequestBody.fromBytes("b2v1".getBytes()));